@@ -10,9 +10,11 @@
     "@fortawesome/fontawesome-free": "^5.11.2",
     "papaparse": "^5.1.1",
     "prop-types": "^15.7.2",
-    "react": "^16.11.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-addons-create-fragment": "^15.6.2",
     "react-markdown": "^4.2.2",
+    "style-loader": "^1.1.3",
     "sweetalert2": "^8.18.6"
   },
   "devDependencies": {
@@ -21,6 +23,9 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
     "@babel/preset-env": "^7.6.3",
     "@babel/preset-react": "^7.6.3",
+    "c3": "^0.7.15",
+    "d3": "^5.15.0",
+    "css-loader": "^3.4.2",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.5",
     "eslint": "^5.16.0",
