@@ -26,10 +26,12 @@ class SiteIDGenerator extends IdentifierGenerator
 {
     private const LENGTH = 4;
 
-    /* Either 'PSCID' or 'ExternalID' */
+    /**
+     * Either 'PSCID' or 'ExternalID'
+     *
+     * @var string
+     */
     protected $kind;
-    protected $siteAlias;
-    protected $projectAlias;
 
     /**
      * Creates a new instance of a SiteIDGenerator to create either PSCIDs or
@@ -47,30 +49,16 @@ class SiteIDGenerator extends IdentifierGenerator
         // Read config settings from project/config.xml to retrieve the
         // alphabet, length, and generation method (sequential or random) used
         // to create new IDs.
-        $this->generationMethod = $this->_getIDSetting('generation');
-        $this->length           = $this->_getIDSetting('length') ?? self::LENGTH;
-        $this->alphabet         = $this->_getIDSetting('alphabet');
-        // Initialize minimum and maximum allowed values for IDs. Set the values
-        // to the lowest/highest character in $alphabet repeated $length times
-        // if the min or max is not configured in project/config.xml
-        $this->maxValue = $this->_getIDSetting('max') ??
-            str_repeat(strval($this->alphabet[0]), intval($this->length));
-        if (!is_array($this->alphabet)) {
-            throw new \ConfigurationException(
-                'Expecting variable $alphabet to be an array but got '
-                . gettype($this->alphabet)
-            );
-        }
-        $this->maxValue = $this->_getIDSetting('max') ??
-            str_repeat(
-                strval($this->alphabet[count($this->alphabet) - 1]),
-                intval($this->length)
-            );
+        $this->generationMethod = $this->_getGeneration();
+        $this->length           = $this->_getLength();
+        $this->alphabet         = $this->_getAlphabet();
+        $this->minValue         = $this->_getMinValue();
+        $this->maxValue         = $this->_getMaxValue();
 
         $this->siteAlias    = $siteAlias;
         $this->projectAlias = $projectAlias;
 
-        $this->prefix = $this->_getIDSetting('prefix');
+        $this->prefix = $this->_getPrefix();
         $this->validate();
     }
 
@@ -150,7 +138,7 @@ class SiteIDGenerator extends IdentifierGenerator
             return array();
         }
         // Filter out non-numeric ids if using a numeric alphabet.
-        if ($this->alphabet === range('0', '9')) {
+        if (empty(array_diff($this->alphabet, range('0', '9')))) {
             return array_filter($ids, 'is_numeric');
         }
         return $ids;
@@ -160,9 +148,9 @@ class SiteIDGenerator extends IdentifierGenerator
      * settings relating to the PSCID structure.
      *
      * @param string $setting One of: 'generation', 'length', 'alphabet',
-     *                        'length', 'min', 'max'.
+     *                        'min', 'max'.
      *
-     * @return array|int|string|null
+     * @return array<int,int|float>|string|null
      */
     private function _getIDSetting(
         string $setting
@@ -233,19 +221,19 @@ class SiteIDGenerator extends IdentifierGenerator
                 );
             }
         }
-        // Min, max, and length values should be returned as integers or as
+        // The remaining values - min, max, and length - should be returned as
         // null if they are not set.
-        return is_null($seqValue) ? $seqValue: intval($seqValue);
+        return is_null($seqValue) ? $seqValue: $seqValue;
     }
     /**
      * Iterate over each 'seq' value and return its setting if its value is
      * configured. Do error handling to make sure that there is exactly one
      * value corresponding to the requested setting.
      *
-     * @param array  $idStructure Settings concerning ID structure extracted
-     *                            from project/config.sml
-     * @param string $setting     The name of the variable for which we want the
-     *                            value.
+     * @param array<array> $idStructure Settings concerning ID structure
+     *                                  extracted from project/config.sml
+     * @param string       $setting     The name of the variable for which we
+     *                                  want the value.
      *
      * @throws \ConfigurationException
      *
@@ -304,12 +292,12 @@ class SiteIDGenerator extends IdentifierGenerator
      * Traverse the $idStructure array and collect all values that exist
      * for $setting.
      *
-     * @param array  $idStructure Settings concerning ID structure extracted
-     *                            from project/config.xml
-     * @param string $setting     The name of the variable for which we want the
-     *                            value.
+     * @param array<array> $idStructure Settings concerning ID structure
+     *                                  extracted from project/config.xml
+     * @param string       $setting     The name of the variable for which
+     *                                  we want the value.
      *
-     * @return array The value(s) corresponding to $setting.
+     * @return array<int,mixed> The value(s) corresponding to $setting.
      */
     private static function _getSeqAttribute(
         array $idStructure,
@@ -322,5 +310,87 @@ class SiteIDGenerator extends IdentifierGenerator
             }
         }
         return $seqAttributes;
+    }
+
+    /**
+     * Initializes the alphabet property.
+     *
+     * @return array<int,int|float|string>
+     */
+    private function _getAlphabet(): array
+    {
+        $alphabet = $this->_getIDSetting('alphabet');
+        if (!is_array($alphabet)) {
+            throw new \ConfigurationException(
+                'Expecting variable $alphabet to be an array but got '
+                . gettype($this->alphabet)
+            );
+        }
+        return $alphabet;
+    }
+
+    /**
+     * Returns the value to be used for the length property.
+     *
+     * @return int
+     */
+    private function _getLength()
+    {
+        $length = $this->_getIDSetting('length') ?? self::LENGTH;
+        return intval($length);
+    }
+
+    /**
+     * Returns the value for the alphabet property.
+     *
+     * @return string
+     */
+    private function _getGeneration(): string
+    {
+        $generation = $this->_getIDSetting('generation');
+        if (!in_array($generation, ['sequential', 'random'], true)) {
+            throw new \ConfigurationException(
+                'Generation method must be either `sequential` or `random`.'
+            );
+        }
+        return $generation;
+    }
+
+    /**
+     * Returns the prefix to be use for the prefix property.
+     *
+     * @return string
+     */
+    private function _getPrefix(): string
+    {
+        return strval($this->_getIDSetting('prefix'));
+    }
+    /**
+     * Returns the minimum value for the identifier.
+     *
+     * @return string
+     */
+    private function _getMinValue(): string
+    {
+        return strval(
+            $this->_getIDSetting('min') ??
+            str_repeat(strval($this->alphabet[0]), intval($this->length))
+        );
+    }
+
+    /**
+     * Returns the maximum value for the identifier.
+     *
+     * @return string
+     */
+    private function _getMaxValue(): string
+    {
+        return strval(
+            $this->_getIDSetting('max') ??
+            str_repeat(
+                strval($this->alphabet[count($this->alphabet) - 1]),
+                intval($this->length)
+            )
+        );
     }
 }
