@@ -22,7 +22,7 @@ namespace LORIS\StudyEntities\Candidate;
  * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
  * @link     https://www.github.com/aces/Loris/
  */
-class CandID extends ValidatableIdentifier
+class CandID extends ValidatableIdentifier implements \JsonSerializable
 {
     /*
      * The minimum allowed value for valid CandIDs. Origin unclear but
@@ -76,6 +76,19 @@ class CandID extends ValidatableIdentifier
      * @return string That CandID's value
      */
     public function __toString(): string
+    {
+        return $this->value;
+    }
+
+    /**
+     * Specify data which should be serialized to JSON.
+     * Returns data which can be serialized by json_encode(), which is a value of
+     * any type other than a resource.
+     *
+     * @see https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     * @return mixed
+     */
+    public function jsonSerialize()
     {
         return $this->value;
     }
